Fixes #34 Open or Save in hubiC doesn't work and #35 Error when loading hubic drive tree

+++ KeeAnywhere/StorageProviders/HubiC/SwiftContainer.cs
@@ -12,6 +12,6 @@ namespace KeeAnywhere.StorageProviders.HubiC
         public int Count { get; set; }
 
         [DataMember(Name = "bytes")]
-        public int Bytes { get; set; }
+        public ulong Bytes { get; set; }
     }
 }
